Monkey patch disable referential integrity during specs.
This is necessary due to the introduction of truncation for database cleaning
in some specs. AR's disable_referential_integrity method for PostgreSQL
requires superuser access if the database has any FKs in it (catch-22!). This
is not an issue in development tests (db:test:prepare strips out FKs), but is
a problem in CI (whose database is built with db:migrate).