Attempt to provoke #1634.
Although the live system is getting validation failures related to date and
decimal formatting, none of these specs are failing. At least part of it, I
think, is due to ActiveSupport messing with the default serialization for
BigDecimal, making the values valid in the tests. (AS core-ext is not loaded
in the actual ETL environment.)